feat(31433): Cria status demonstrativo financeiro e relação de bens
Altera o modelo do arquivos de demonstrativo financeiro e relação de
bens para incluir campo de status que poderá ser "em processamento" ou
"concluido".
