+++ sme_ptrf_apps/core/models/relacao_bens.py
@@ -5,6 +5,20 @@ from sme_ptrf_apps.core.models_abstracts import ModeloBase
 
 
 class RelacaoBens(ModeloBase):
+    # Status Choice
+    STATUS_EM_PROCESSAMENTO = 'EM_PROCESSAMENTO'
+    STATUS_CONCLUIDO = 'CONCLUIDO'
+
+    STATUS_NOMES = {
+        STATUS_EM_PROCESSAMENTO: 'Em processamento',
+        STATUS_CONCLUIDO: 'Geração concluída',
+    }
+
+    STATUS_CHOICES = (
+        (STATUS_EM_PROCESSAMENTO, STATUS_NOMES[STATUS_EM_PROCESSAMENTO]),
+        (STATUS_CONCLUIDO, STATUS_NOMES[STATUS_CONCLUIDO]),
+    )
+
     arquivo = models.FileField(blank=True, null=True)
 
     conta_associacao = models.ForeignKey('ContaAssociacao', on_delete=models.PROTECT,
@@ -13,12 +27,23 @@ class RelacaoBens(ModeloBase):
     prestacao_conta = models.ForeignKey('PrestacaoConta', on_delete=models.CASCADE,
                                         related_name='relacoes_de_bens_da_prestacao', blank=True, null=True)
 
+    status = models.CharField(
+        'status',
+        max_length=20,
+        choices=STATUS_CHOICES,
+        default=STATUS_EM_PROCESSAMENTO
+    )
+
     class Meta:
         verbose_name = 'Relação de bens'
         verbose_name_plural = '09.3) Relações de bens'
 
     def __str__(self):
         return f"Documento gerado dia {self.criado_em.strftime('%d/%m/%Y %H:%M')}"
+
+    def arquivo_concluido(self):
+        self.status = self.STATUS_CONCLUIDO
+        self.save()
 
 
 @receiver(models.signals.post_delete, sender=RelacaoBens)
