+++ sme_ptrf_apps/core/services/prestacao_contas_services.py
@@ -3,10 +3,10 @@ import logging
 from django.db import transaction
 from django.db.models import Q
 
-from .demonstrativo_financeiro_novo import gerar_arquivo_demonstrativo_financeiro_novo
+from .demonstrativo_financeiro import gerar_arquivo_demonstrativo_financeiro_novo
 from ..models import PrestacaoConta, AcaoAssociacao, FechamentoPeriodo, ContaAssociacao, Associacao
 from ..services import info_acoes_associacao_no_periodo
-from ..services.demonstrativo_financeiro import gerar_arquivo_demonstrativo_financeiro
+from ..services.demonstrativo_financeiro_old import gerar_arquivo_demonstrativo_financeiro
 from ..services.relacao_bens import gerar_arquivo_relacao_de_bens
 from ..services.processos_services import get_processo_sei_da_prestacao
 from ...despesas.models import RateioDespesa
@@ -71,12 +71,8 @@ def _criar_documentos(acoes, contas, periodo, prestacao):
         gerar_arquivo_relacao_de_bens(periodo=periodo, conta_associacao=conta, prestacao=prestacao)
 
         logger.info(f'Gerando demonstrativo financeiro da conta {conta}.')
-        gerar_arquivo_demonstrativo_financeiro_novo(acoes=acoes, periodo=periodo, conta_associacao=conta, prestacao=prestacao)
-
-        # for acao in acoes:
-        #     logger.info(f'Gerando demonstrativo financeiro da ação {acao} e conta {conta}.')
-        #     gerar_arquivo_demonstrativo_financeiro(periodo=periodo, conta_associacao=conta, acao_associacao=acao,
-        #                                            prestacao=prestacao)
+        gerar_arquivo_demonstrativo_financeiro_novo(acoes=acoes, periodo=periodo, conta_associacao=conta,
+                                                    prestacao=prestacao)
 
 
 def reabrir_prestacao_de_contas(prestacao_contas_uuid):
